scripts/text_enc_gen/requirements.txt
freetype-py>=2.0.0
